backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
email-validator
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiogram==3.3.0
prometheus-client==0.19.0
# Monitoring & Error Tracking
sentry-sdk[fastapi]==1.40.0
# Redis async client for caching and health checks
redis==5.0.1
pywebpush==1.14.0
pandas==2.2.0
openpyxl==3.1.2
user-agents==2.2.0
openai==1.10.0
# web3==6.15.0
stripe==7.1.0
qrcode>=7.4.2

# Code Quality
mypy==1.8.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
types-redis
types-python-jose
types-requests
# sqlalchemy-stubs is included in sqlalchemy 2.0+ but wouldn't hurt if we need specific types, 
# though usually not needed for modern alchemy. I'll skip it to avoid conflicts.

